requirements.txt
langchain
bs4
python-dotenv
